+++ Editor/ShaderAnalyzer.cs
@@ -319,50 +319,62 @@ namespace d4rkpl4y3r.AvatarOptimizer
             return -1;
         }
 
-        public static List<(string name, bool notEquals, float value)> ParseIfexConditions(string line)
+        public enum IfexConditionType
         {
-            var conditions = new List<(string name, bool notEquals, float value)>();
+            Equals,
+            NotEquals,
+            IsNotAnimated,
+            IsAnimated
+        }
+        public static List<(string propertyName, IfexConditionType conditionType, float value)> ParseIfexConditions(string line)
+        {
+            var conditions = new List<(string propertyName, IfexConditionType conditionType, float value)>();
             int index = 5;
-            while (index < line.Length) {
-                while (index < line.Length && char.IsWhiteSpace(line[index]))
-                    index++;
+            void SkipWhitespace() { while (index < line.Length && char.IsWhiteSpace(line[index])) index++; }
+            while (index < line.Length)
+            {
+                SkipWhitespace();
                 if (index == line.Length)
                     break;
-                if (line[index] == '&' && line[index + 1] == '&') {
+                if (line[index] == '&' && line[index + 1] == '&')
                     index += 2;
-                }
-                while (index < line.Length && char.IsWhiteSpace(line[index]))
-                    index++;
+                SkipWhitespace();
                 var name = ShaderAnalyzer.ParseIdentifierAndTrailingWhitespace(line, ref index);
-                if (name == null) {
+                if (name == null)
                     return null;
+                if (name == "isNotAnimated" || name == "isAnimated")
+                {
+                    var conditionType = name == "isNotAnimated" ? IfexConditionType.IsNotAnimated : IfexConditionType.IsAnimated;
+                    if (line[index++] != '(')
+                        return null;
+                    SkipWhitespace();
+                    name = ShaderAnalyzer.ParseIdentifierAndTrailingWhitespace(line, ref index);
+                    if (name == null)
+                        return null;
+                    if (line[index++] != ')')
+                        return null;
+                    conditions.Add((name, conditionType, 0f));
+                    continue;
                 }
-                while (index < line.Length && char.IsWhiteSpace(line[index]))
-                    index++;
-                if (index == line.Length) {
+                SkipWhitespace();
+                if (index == line.Length)
                     return null;
-                }
                 bool notEquals = line[index] == '!';
-                if ((line[index] != '!' && line[index] != '=') || line[index + 1] != '=') {
+                if ((line[index] != '!' && line[index] != '=') || line[index + 1] != '=')
                     return null;
-                }
                 index += 2;
-                while (index < line.Length && char.IsWhiteSpace(line[index]))
-                    index++;
-                if (index == line.Length) {
+                SkipWhitespace();
+                if (index == line.Length)
                     return null;
-                }
                 int valueStart = index;
                 int valueEnd = index;
-                while (valueEnd < line.Length && char.IsDigit(line[valueEnd])) {
+                while (valueEnd < line.Length && char.IsDigit(line[valueEnd]))
                     valueEnd++;
-                }
-                if (valueStart == valueEnd) {
+                if (valueStart == valueEnd)
                     return null;
-                }
                 index = valueEnd;
                 var value = float.Parse(line.Substring(valueStart, valueEnd - valueStart));
-                conditions.Add((name, notEquals, value));
+                conditions.Add((name, notEquals ? IfexConditionType.NotEquals : IfexConditionType.Equals, value));
             }
             return conditions;
         }
@@ -499,7 +511,7 @@ namespace d4rkpl4y3r.AvatarOptimizer
                         processedLines.Add(ifexLine);
                         var conditions = ParseIfexConditions(ifexLine);
                         if (conditions != null) {
-                            conditions.ForEach(p => parsedShader.ifexParameters.Add(p.name));
+                            conditions.ForEach(p => parsedShader.ifexParameters.Add(p.propertyName));
                         } else {
                             parsedShader.unableToParseIfexStatements.Add(trimmedLine);
                         }
@@ -3141,24 +3153,48 @@ namespace d4rkpl4y3r.AvatarOptimizer
             var outputString = $"// #ifex ";
             var firstCondition = true;
             foreach (var condition in conditions) {
-                var name = condition.name;
-                var notEquals = condition.notEquals;
-                var compValue = condition.value;
-                if (!staticPropertyValues.TryGetValue(name, out var valueString)) {
-                    lineIndex++;
-                    debugOutput?.Add($"// #ifex {name} not found in static properties");
-                    ParseAndEvaluateIfex(lines, ref lineIndex, debugOutput);
-                    return;
+                var propertyName = condition.propertyName;
+                bool isStaticValue = staticPropertyValues.TryGetValue(propertyName, out var staticValue);
+                bool isAnimated = animatedPropertyValues.ContainsKey(propertyName);
+                bool isArrayProperty = arrayPropertyValues.ContainsKey(propertyName);
+                bool isValueConstant = isStaticValue && !isAnimated && !isArrayProperty;
+                outputString += $"{(firstCondition ? "" : " && ")}";
+                switch (condition.conditionType)
+                {
+                    case ShaderAnalyzer.IfexConditionType.IsAnimated:
+                    case ShaderAnalyzer.IfexConditionType.IsNotAnimated:
+                        bool isAnimatedCheck = condition.conditionType == ShaderAnalyzer.IfexConditionType.IsAnimated;
+                        outputString += $"{(isAnimatedCheck ? "isAnimated" : "isNotAnimated")}({propertyName})({isStaticValue},{isAnimated},{isArrayProperty})";
+                        if (isValueConstant == isAnimatedCheck || propertyName == "_AlphaToCoverage")
+                        {
+                            lineIndex++;
+                            debugOutput?.Add(outputString + ", FALSE" + (propertyName == "_AlphaToCoverage" ? " (special case)" : ""));
+                            ParseAndEvaluateIfex(lines, ref lineIndex, debugOutput);
+                            return;
+                        }
+                        break;
+                    case ShaderAnalyzer.IfexConditionType.Equals:
+                    case ShaderAnalyzer.IfexConditionType.NotEquals:
+                        if (!isValueConstant)
+                        {
+                            lineIndex++;
+                            debugOutput?.Add($"// #ifex {propertyName} not a constant value");
+                            ParseAndEvaluateIfex(lines, ref lineIndex, debugOutput);
+                            return;
+                        }
+                        var value = float.Parse(staticValue);
+                        var notEquals = condition.conditionType == ShaderAnalyzer.IfexConditionType.NotEquals;
+                        outputString += $"{propertyName}({value}) {(notEquals ? '!' : '=')}= {condition.value}";
+                        if ((condition.value == value) == notEquals)
+                        {
+                            lineIndex++;
+                            debugOutput?.Add(outputString + ", FALSE");
+                            ParseAndEvaluateIfex(lines, ref lineIndex, debugOutput);
+                            return;
+                        }
+                        break;
                 }
-                var value = float.Parse(valueString);
-                outputString += $"{(firstCondition ? "" : " && ")}{name}({value}) {(notEquals ? '!' : '=')}= {compValue}";
                 firstCondition = false;
-                if ((compValue != value) ^ notEquals) {
-                    lineIndex++;
-                    debugOutput?.Add(outputString + ", FALSE");
-                    ParseAndEvaluateIfex(lines, ref lineIndex, debugOutput);
-                    return;
-                }
             }
 
             // skip all code until matching #endex